pyproject.toml
[tool.poetry]
name = "ML API Server"
version = "0.1.0"
description = "ML service for Bidhive"
authors = ["Danil Dvorchuk <dvorchyk.d.dev@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.65.2"
requests = "^2.25.1"
uvicorn = "^0.15.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pydantic = "^1.10.4"
python-dotenv = "^0.18.0"
python-multipart = "^0.0.5"
bcrypt = "^3.2.0"
email-validator = "^1.2.1"
anyio = "^3.6.1"
setuptools = "^67.3.2"
moto = {extras = ["server"], version = "^4.1.4"}
boto3 = "~1.24.59"
aioboto3 = "^10.4.0"
celery = "^5.2.7"
redis = "^4.5.1"
sqlalchemy = "^2.0.4"
psycopg2-binary = "^2.9.5"
asyncpg = "^0.27.0"
alembic = "^1.9.4"
invoke = "^2.0.0"
ipython = "^8.10.0"
python-docx = "^0.8.11"
docx2txt = "^0.8"
pandas = "^1.5.3"
s3fs = "^2023.1.0"
openai = "^0.27.0"
tiktoken = "^0.3.0"
matplotlib = "^3.7.1"
plotly = "^5.13.1"
scipy = "^1.10.1"
scikit-learn = "^1.2.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"
pytest = "^7.1.3"
ipython = "^8.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
